Add files via upload
A Simple Calculator is designed using Tkinter Module in Python.Tkinter is the standard GUI (Graphical User Interface) toolkit for Python. It provides a simple and easy-to-use interface for building desktop applications with graphical components such as windows, buttons, textboxes, and more.This Calculator uses the features of Tkinter library and performs arithmetic operations such as addition, subtraction, multiplication, and division.